+++ CHANGELOG.md
@@ -32,7 +32,7 @@
 	- `IPXCache` - `PP\Lib\Cache\CacheInterface`
 - `PXAbstractCronRun` — deprecated, использовать `PP\Cron\AbstractCron`
 - `PXCronRule` — deprecated, использовать `PP\Cron\CronRule`
-- `PXAbstractPlugin` — deprecated, использовать `PP\Plugin\PluginAbstract`
+- `PXAbstractPlugin` — deprecated, использовать `PP\Plugin\AbstractPlugin`
 - `NLPGSQLDatabase` — deprecated, использовать `PP\Lib\Database\DatabaseSqlAbstract`
 - `charcheck` - убран из всех форм, javascript и php кода
 - Работа с файлом `database.ini` удалена, настройки подключения к базе хранятся в environment