+++ CHANGELOG.md
@@ -30,10 +30,10 @@
 	- `ObjectCache` - `PP\Lib\Cache\ObjectCache`
 	- `PXCache*` - `PP\Lib\Cache\Driver\*`
 	- `IPXCache` - `PP\Lib\Cache\CacheInterface`
-- `PXAbstractCronRun` — deprecated, использовать `PP\Cron\CronAbstract`
+- `PXAbstractCronRun` — deprecated, использовать `PP\Cron\AbstractCron`
 - `PXCronRule` — deprecated, использовать `PP\Cron\CronRule`
-- `PXAbstractPlugin` — deprecated, использовать `PP\Plugin\PluginAbstract`
-- `NLPGSQLDatabase` — deprecated, использовать `PP\Lib\Database\DatabaseSqlAbstract`
+- `PXAbstractPlugin` — deprecated, использовать `PP\Plugin\AbstractPlugin`
+- `NLPGSQLDatabase` — deprecated, использовать `PP\Lib\Database\AbstractSqlDatabase`
 - `charcheck` - убран из всех форм, javascript и php кода
 - Работа с файлом `database.ini` удалена, настройки подключения к базе хранятся в environment
 	- [Документация](docs/configuration.md)